+++ tools/unit_test.h
@@ -14,10 +14,8 @@
 #ifdef FRESA_ENABLE_TESTS
 
 #include "std_types.h"
+#include "source_loc.h"
 #include "log.h"
-#ifdef __cpp_lib_source_location
-    #include <source_location>
-#endif
 
 namespace fresa
 {
@@ -31,34 +29,6 @@ namespace fresa
 
     namespace detail
     {
-        //* source location
-        //      provides the name of the file and the line number of the function calling it
-        //      standard in c++20, but compiler support is lacking, so an alternative is also provided
-        #ifdef __cpp_lib_source_location
-            using source_location = std::source_location;
-        #else
-            struct source_location
-            {
-                static constexpr auto current(
-                    #if (__has_builtin(__builtin_FILE) && __has_builtin(__builtin_LINE))
-                        const char *file = __builtin_FILE(),
-                        int line = __builtin_LINE()
-                    #else
-                        const char *file = "no file",
-                        int line = 0
-                    #endif
-                ) noexcept {
-                    return source_location{file, line};
-                }
-
-                constexpr auto file_name() const noexcept { return file_; }
-                constexpr auto line() const noexcept { return line_; }
-
-                const char* file_;
-                int line_;
-            };
-        #endif
-
         //* test objects
         //      useful structures to store data for the test system
         namespace test_objects